updated data fields to float

+++ app/models/workout.rb
@@ -2,7 +2,7 @@ class Workout
   include Mongoid::Document
   field :workout_date, :type => Date
   field :description, :type => String
-  field :power, :type => Decimal
+  field :power, :type => Float
   field :crossfit_workout_id, :type => Integer
-  field :mileage, type: Decimal
+  field :mileage, type: Float
 end
